Major Site Revamp
This is probably a little more major than is strictly necessary, *but*
it does set up a pretty good framework for building in the future.

I added Bootstrap (actually a Bootswatch theme) so it's pretty without
having to spend too much time making it look good. The benefit of this
is that just about everything you'll want to add in the future will look
good with this. I'm getting Bootstrap (css and js) from a CDN. Note that
the version of jQuery I'm using isn't compatiable with old versions of
Internet Explorer (but, since this is an OS X app, I doubt if that's a
big problem users will face).

I also moved just about everything that I could out into includes. This
makes it a lot easier to modify things in the future, since you've only
go to do it in one place.

Everything But Docs

As of now, to the best of my knowledge, everything works except the
documentation.

Some of the local jekyll testing is kind of funky. I remember this being
a problem when working on my personal site, but I don't really remember
the resolution. I remember a *lot* of commits and pushing of code trying
to make it work correctly.